Refactored status cli into a different file
